remove all vite lib exports for specs
