+++ code/ext/ext.lib.reactflow/vite.config.ts
@@ -5,12 +5,7 @@ export const tsconfig = Config.ts((e) => {
 });
 
 export default Config.vite(import.meta.url, (e) => {
-  e.lib({
-    entry: {
-      index: 'src/index.ts',
-      specs: 'src/test.ui/entry.Specs.ts',
-    },
-  });
+  e.lib({ entry: { index: 'src/index.ts' } });
   e.target('web');
   e.externalDependency(e.ctx.deps.map((d) => d.name));
 });
