+++ code/ext/ext.lib.stripe/src/index.ts
@@ -1,1 +1,6 @@
 export { Pkg } from './index.pkg';
+
+/**
+ * Dev
+ */
+export { Specs } from './test.ui/entry.Specs';
