+++ code/ext/ext.lib.stripe/vite.config.ts
@@ -8,7 +8,6 @@ export default Config.vite(import.meta.url, (e) => {
   e.lib({
     entry: {
       index: 'src/index.ts',
-      specs: 'src/test.ui/entry.Specs.ts',
     },
   });
   e.target('web');
