+++ code/sys.ui/react.dev/src/index.ts
@@ -1,5 +1,8 @@
 export { Pkg } from './index.pkg';
 
+/**
+ * Library
+ */
 export { BADGES, DevWrangle, Is } from './common';
 export { DevKeyboard } from './f.Keyboard';
 export { Spec } from './f.Spec';
@@ -10,3 +13,8 @@ export { useRubberband } from './ui.use/useRubberband';
 
 export { Dev } from './Dev';
 export { render } from './Dev.render';
+
+/**
+ * Dev
+ */
+export { Specs, ModuleSpecs, SampleSpecs } from './test.ui/entry.Specs';
